ship-service/requirements.txt
django
djangorestframework
PyMySQL


cryptography
